// worker/src/types/env.ts
export type AppBindings = {
  DB: D1Database;
  KV: KVNamespace;
  R2: R2Bucket;
  DOMAINS: string;
  PREFIX: string;
  JWT_SECRET: string;
  ADMIN_PASSWORDS: string;
  ENABLE_ADDRESS_PASSWORD: string;
  DISABLE_CUSTOM_ADDRESS_NAME: string;
  CREATE_ADDRESS_DEFAULT_DOMAIN_FIRST: string;
  DEFAULT_DOMAINS: string;
  ANNOUNCEMENT: string;
  ALWAYS_SHOW_ANNOUNCEMENT: string;
  TURNSTILE_SECRET: string;
  DEBUG_MODE: string;
  MAX_ATTACHMENT_SIZE: string;
  REMOVE_ALL_ATTACHMENT: string;
  REMOVE_EXCEED_SIZE_ATTACHMENT: string;
  FORWARD_ADDRESS_LIST: string;
  WEBHOOK_URL: string;
  TELEGRAM_BOT_TOKEN: string;
  TELEGRAM_BOT_WEBHOOK_SECRET: string;
  DKIM_PRIVATE_KEY: string;
  RESEND_API_KEY: string;
  SMTP_HOST: string;
  SMTP_PORT: string;
  SMTP_USERNAME: string;
  SMTP_PASSWORD: string;
  S3_REGION: string;
  S3_ENDPOINT: string;
  S3_ACCESS_KEY_ID: string;
  S3_SECRET_ACCESS_KEY: string;
  S3_BUCKET: string;
  OAUTH2_PROVIDERS: string;
  USER_ROLES: string;
  USER_DEFAULT_ROLE: string;
  APP_ORIGINS?: string;
  WEBHOOK_SECRET?: string;
};